Rearrange file to clarify division of responsibility
This commit rearranges the package source to clarify that it really
consists of two parts, which could function independently:
* A Flymake backend that, when nudged, takes errors from the log file,
parses them with AUCTeX, and feeds them to Flymake.  (This could also
be hooked up to AUCTeX's normal way of compiling things, giving an
alternative way to manage errors.)
* A minor mode that continuously compiles using latexmk,
and (optionally) nudges the Flymake backend on a timer.